Naive heightmap test

+++ source/CjClutter.OpenGl/SceneGraph/Face3.cs
@@ -2,8 +2,8 @@ namespace CjClutter.OpenGl.SceneGraph
 {
     public struct Face3
     {
-        public ushort V0;
-        public ushort V1;
-        public ushort V2;
+        public uint V0;
+        public uint V1;
+        public uint V2;
     }
 }
